:art: Better documentation
